requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
strawberry-graphql
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
email-validator
